requirements.txt
pytelegrambotapi==3.7.4
loguru==0.5.3
PyYAML>=5.4.1
terminaltables>=3.1.0
